pyproject.toml
[tool.poetry]
name = "shrugbot"
version = "0.1.0"
description = "Shrug Bot for Telegram"
authors = ["nyuszika7h <nyuszika7h@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = "^13.7"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
